Move rendering triangulator to raytracing module
This triangulator creates arrays specifically designed for the
raytracing module. Since it is custom made for the raytracing
module, it will help with dependencies to make it reside there.
