+++ viskores/rendering/MapperRayTracer.cxx
@@ -23,11 +23,11 @@
 #include <viskores/cont/TryExecute.h>
 
 #include <viskores/rendering/CanvasRayTracer.h>
-#include <viskores/rendering/internal/RunTriangulator.h>
 #include <viskores/rendering/raytracing/Camera.h>
 #include <viskores/rendering/raytracing/Logger.h>
 #include <viskores/rendering/raytracing/RayOperations.h>
 #include <viskores/rendering/raytracing/RayTracer.h>
+#include <viskores/rendering/raytracing/RunTriangulator.h>
 #include <viskores/rendering/raytracing/SphereExtractor.h>
 #include <viskores/rendering/raytracing/SphereIntersector.h>
 #include <viskores/rendering/raytracing/TriangleExtractor.h>
